More major rework, refactor, restructure. Currently works.
