+++ mensio/config.yaml
@@ -1,9 +1,12 @@
 # This file can be used to change inputs to the XML to Excel program.
 
-# Excel path relative to main.py.
+# Excel path relative to config.yaml location (mensio/).
 excel_path: ../data/2025-09_24_Patient_Measurements_VX.xlsx
+
 sheet_title: tavi_mock_1000_with_G  # Name of tab to write in.
-xml_directory: ../data/Datenaus3Mensio/  # Path to the directory with XML files relative to main.py.
+
+# Path to the directory with XML files relative to config.yaml location (mensio/).
+xml_directory: ../data/Datenaus3Mensio/
 
 # The mapping of Excel column numbers to XML tags.
 mapping:
