+++ tests/test_xml2xlsx.py
@@ -1,11 +1,11 @@
 """High level tests."""
 
-from src.main import check_config_data, get_xml_files
-from src.modules.excel_writer import ExcelWriter
+from main import check_config_data, get_xml_files
+from modules.excel_writer import ExcelWriter
 
 
 def test_xml2xlsx():
-    """Test xml2xlsx."""
+    """Test mensio."""
     config_data, xml_directory, excel_path = check_config_data()
     excel_writer: ExcelWriter = ExcelWriter(config_data)
     xml_files: list[str] = get_xml_files(xml_directory)
